# server/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "سوق الرافدين — Rafidain Market API",
    "version": "1.0.0",
    "description": "عقد API لمنصة سوق الرافدين (وساطة متعددة الخدمات: متاجر/مطاعم/فنادق/طيران/مكاتب سفر). مواصفة أساسية (v1) تغطّي المسارات الأساسية بمخططات حقيقية، قابلة للتوسعة لجميع ملفات المسارات الـ29. الأدوار: admin / agent / provider / customer. العملة: الدينار العراقي (IQD).",
    "contact": { "name": "فريق الرافدين", "email": "dev@rafidain.iq" }
  },
  "servers": [
    { "url": "http://localhost:4001/api", "description": "تطوير محلي" },
    { "url": "https://api.rafidain.iq/api", "description": "إنتاج" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "rafidain_session" }
    },
    "schemas": {
      "Error": { "type": "object", "properties": { "message": { "type": "string" }, "code": { "type": "string" } } },
...
      "ProviderSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name_ar": { "type": "string" },
          "service_slug": { "type": "string" },
          "governorate_id": { "type": "integer" },
          "rating": { "type": "number" },
          "rating_count": { "type": "integer" },
          "is_verified": { "type": "boolean" },
          "is_active": { "type": "boolean" },
          "logo": { "type": "string", "nullable": true }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "provider_id": { "type": "integer" },
          "title_ar": { "type": "string" },
          "price": { "type": "number" },
          "discount_percent": { "type": "number" },
          "stock": { "type": "integer" },
          "images": { "type": "array", "items": { "type": "string" } },
          "is_active": { "type": "boolean" }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "kind": { "type": "string", "enum": ["product", "menu_item", "room", "flight", "package"] },
          "item_id": { "type": "integer" },
          "title_ar": { "type": "string" },
          "quantity": { "type": "integer" },
          "unit_price": { "type": "number" }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "order_number": { "type": "string" },
          "customer_id": { "type": "integer" },
          "provider_id": { "type": "integer" },
          "status": { "type": "string", "enum": ["pending", "confirmed", "in_progress", "completed", "cancelled"] },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/OrderItem" } },
          "total_amount": { "type": "number" },
          "commission_amount": { "type": "number" },
          "platform_amount": { "type": "number" },
          "agent_amount": { "type": "number" },
          "provider_amount": { "type": "number" },
          "coupon_code": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "NotificationPreferences": {
        "type": "object",
        "properties": {
          "in_app": { "type": "boolean", "description": "فعالية القناة داخل التطبيق" },
          "push": { "type": "boolean", "description": "فعالية Web Push" },
          "categories": {
            "type": "object",
            "description": "خريطة تفعيل الفئات (order/wallet/promotions/announcement/providers)"
          }
        }
      },
      "Dispute": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "order_id": { "type": "integer" },
          "opened_by": { "type": "integer" },
          "party": { "type": "string" },
          "reason": { "type": "string" },
          "status": { "type": "string", "enum": ["open", "reviewing", "resolved", "rejected"] },
          "resolution": { "type": "string", "nullable": true },
          "refund_amount": { "type": "number" },
          "order_number": { "type": "string" }
        }
      },
      "Invoice": {
        "type": "object",
        "properties": {
          "invoice": { "type": "object", "properties": { "number": { "type": "string" }, "subtotal": { "type": "number" }, "discount_amount": { "type": "number" }, "tax_rate": { "type": "number" }, "tax_amount": { "type": "number" }, "total": { "type": "number" }, "currency": { "type": "string" } } },
          "items": { "type": "array" }
        }
      },
      "PaginatedOrders": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Order" } },
          "meta": { "type": "object", "properties": { "total": { "type": "integer" }, "page": { "type": "integer" }, "limit": { "type": "integer" }, "pages": { "type": "integer" } } }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/public/providers": {
      "get": {
        "summary": "بحث/تصفّح المزوّدين",
        "tags": ["public"],
        "security": [],
        "parameters": [
          { "name": "service", "in": "query", "schema": { "type": "string" } },
          { "name": "governorate_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": { "200": { "description": "قائمة", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProviderSummary" } } } } } }
      }
    },
    "/public/providers/{id}": {
      "get": {
        "summary": "تفاصيل مزوّد",
        "tags": ["public"],
        "security": [],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "المزوّد", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProviderSummary" } } } }, "404": { "description": "غير موجود", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } } }
      }
    },
    "/public/providers/{id}/products": {
      "get": {
        "summary": "منتجات/قائمة مزوّد",
        "tags": ["public"],
        "security": [],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "قائمة", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } } } } } }
      }
    },
    "/public/providers/{id}/reviews": {
      "get": {
        "summary": "تقييمات مزوّد",
        "tags": ["public"],
        "security": [],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "قائمة تقييمات" } }
      }
    },
    "/public/items/{kind}/{itemId}/reviews": {
      "get": {
        "summary": "تقييمات عنصر (product/menu_item/room/flight/package)",
        "tags": ["public"],
        "security": [],
        "parameters": [
          { "name": "kind", "in": "path", "required": true, "schema": { "type": "string", "enum": ["product", "menu_item", "room", "flight", "package"] } },
          { "name": "itemId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "قائمة تقييمات" } }
      }
    },
    "/orders": {
      "get": {
        "summary": "قائمة الطلبات (مُرشَّحة حسب الدور)",
        "tags": ["orders"],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": { "200": { "description": "قائمة", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaginatedOrders" } } } } }
      },
      "post": {
        "summary": "إنشاء طلب وساطة",
        "tags": ["orders"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["provider_id", "items"], "properties": { "provider_id": { "type": "integer" }, "items": { "type": "array", "items": { "$ref": "#/components/schemas/OrderItem" } }, "coupon_code": { "type": "string" } } } } } },
        "responses": { "201": { "description": "أنشئ", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } } }, "400": { "description": "رصيد/بيانات غير صالحة", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } } }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "تفاصيل طلب",
        "tags": ["orders"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "الطلب", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } } }, "404": { "description": "غير موجود", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } } }
      }
    },
    "/orders/{id}/status": {
      "put": {
        "summary": "تغيير حالة الطلب (صلاحيات حسب الدور)",
        "tags": ["orders"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["status"], "properties": { "status": { "type": "string", "enum": ["pending", "confirmed", "in_progress", "completed", "cancelled"] } } } } } },
        "responses": { "200": { "description": "محدَّث", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } } }, "403": { "description": "صلاحية", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } } }
      }
    },
    "/orders/{id}/invoice": {
      "get": { "summary": "فاتورة طلب (JSON) — بنطاق الصلاحية", "tags": ["orders"], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "فاتورة", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invoice" } } } }, "403": { "description": "صلاحية" } } }
    },
    "/orders/{id}/disputes": {
      "post": { "summary": "فتح نزاع على طلب مكتمل/ملغي", "tags": ["disputes"], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["reason"], "properties": { "reason": { "type": "string" } } } } } }, "responses": { "201": { "description": "فُتح", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Dispute" } } } }, "422": { "description": "طلب غير مؤهل" } } }
    },
...
    "/payments/charge": {
      "post": { "summary": "بدء عملية دفع عبر بوابة (تُرد 501 ما لم تكن مفعّلة)", "tags": ["payments"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["gateway_code", "amount"], "properties": { "gateway_code": { "type": "string" }, "amount": { "type": "number" } } } } } }, "responses": { "200": { "description": "جاهزة للاستخدام بعد التفعيل", "content": { "application/json": { "schema": { "type": "object", "properties": { "allowed": { "type": "boolean" }, "gateway_code": { "type": "string" } } } } } }, "501": { "description": "بوابة غير مفعّلة" } } }
    },
...
    "/payments/webhook/{gateway}": {
      "post": { "summary": "نقطة وصل موحّدة لبوابة (تُرد 501 ما لم تكن مفعّلة)", "tags": ["payments"], "parameters": [{ "name": "gateway", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "501": { "description": "بوابة غير مفعّلة" } } }
    },
...
    "/notifications/{id}/read": {
      "post": { "summary": "تعليم إشعار كمقروء", "tags": ["notifications"], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "تم" } } }
    },
    "/notifications/preferences": {
      "get": { "summary": "تفضيلات الإشعارات للمصادَق", "tags": ["notifications"], "responses": { "200": { "description": "تفضيلات (قنوات وفئات)" } } },
      "put": { "summary": "تحديث تفضيلات الإشعارات (in_app/فئات)", "tags": ["notifications"], "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NotificationPreferences" } } } }, "responses": { "200": { "description": "حُفظت" }, "400": { "description": "فئة غير معروفة" } } }
    },
...
  }
}